+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 This list is not intended to be all-encompassing - it will document major and breaking API 
 changes with their rationale when appropriate. Given version `A.B.C.D`, breaking changes are to be expected in version number increments where changes in the `A` or `B` sections:
 
+### v5.17.0.2
+- **http4k-connect-ai-openai** - Choices are not optional in conversation completion.
+- **http4k-connect-ai-langchain** - Fix streaming for OpenAiChatLanguageModel. 
+
 ### v5.17.0.1
 - **http4k-connect-ai-langchain** - Added support for System messages in Ollama models